+++ cg_lims/EPPs/files/base.py
@@ -3,10 +3,10 @@ import click
 from cg_lims.EPPs.files.barcode_tubes import make_barcode_csv
 from cg_lims.EPPs.files.csv_for_kapa_truble_shooting.csv_for_kapa_debug import trouble_shoot_kapa
 from cg_lims.EPPs.files.femtopulse_csv import make_femtopulse_csv
-from cg_lims.EPPs.files.file_to_udf import csv_well_to_udf
 from cg_lims.EPPs.files.hamilton.base import hamilton
-from cg_lims.EPPs.files.illumina_xml_to_udf import parse_run_parameters
-from cg_lims.EPPs.files.ont_json_to_udf import parse_ont_report
+from cg_lims.EPPs.files.parsers.file_to_udf import csv_well_to_udf
+from cg_lims.EPPs.files.parsers.illumina_xml_to_udf import parse_run_parameters
+from cg_lims.EPPs.files.parsers.ont_json_to_udf import parse_ont_report
 from cg_lims.EPPs.files.placement_map.make_96well_placement_map import placement_map
 from cg_lims.EPPs.files.pooling_map.make_pooling_map import pool_map
 from cg_lims.EPPs.files.sample_sheet.create_ont_sample_sheet import create_ont_sample_sheet
